Defining all routes, and state management/ network is now at stake
