+++ membership_card/.packages
@@ -1,11 +1,13 @@
-# Generated by pub on 2019-10-31 23:00:53.608022.
+# Generated by pub on 2019-11-01 09:42:36.867317.
 async:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/async-2.3.0/lib/
 boolean_selector:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/boolean_selector-1.0.5/lib/
 charcode:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/charcode-1.1.2/lib/
 collection:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/collection-1.14.11/lib/
 cupertino_icons:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/cupertino_icons-0.1.2/lib/
+dio:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/dio-3.0.4/lib/
 flutter:file:///C:/flutter/packages/flutter/lib/
 flutter_test:file:///C:/flutter/packages/flutter_test/lib/
+http_parser:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/http_parser-3.1.3/lib/
 matcher:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/matcher-0.12.5/lib/
 meta:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/meta-1.1.7/lib/
 path:file:///C:/flutter/.pub-cache/hosted/pub.flutter-io.cn/path-1.6.4/lib/
