+++ membership_card/lib/model/card.dart
@@ -1,4 +1,6 @@
+import 'package:flutter/material.dart';
+
 /// This is the class defining all the card info
-class CardInfo {
+class CardInfo extends ChangeNotifier{
   //Todo: finishing all the card info here
 }
