+++ membership_card/lib/pages/add_cards.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 /// This is the Add_Card Page Class especially for adding one card
