+++ membership_card/pubspec.yaml
@@ -22,6 +22,8 @@ dependencies:
 
   provider: ^3.0.1+1
 
+  dio: ^3.0.0
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
